+++ __tests__/__configs__/jest.config.js
@@ -1,8 +1,11 @@
 const path = require('path');
 
 module.exports = {
-  rootDir: path.join(__dirname, '../'),
+  collectCoverage: true,
+  testMatch: ['**/__tests__/**/*.test.js?(x)'],
+  coveragePathIgnorePatterns: ['/node_modules/', '/__configs__/', '/src/', '/view/'],
+  testPathIgnorePatterns: ['/node_modules/', '/__configs__/', '/src/', '/view/'],
+  rootDir: path.join(__dirname, '../../'),
   globals: { __ROOT__: path.join(__dirname, '../../') },
-  testPathIgnorePatterns: ['/node_modules/', '/__configs__/'],
   setupFiles: [path.join(__dirname, 'env.config.js')],
 };